Update LICENSE.md
http://bkhataloansinbangalore.blogspot.in/

Khata is critical when you apply for any permit of building or for exchange, applying for credit from any banks or monetary foundation. Khata is a record which comprise every one of the subtle elements of property like name of proprietor, size of structures, area of property and every other point of interest that documents property charge. Taking after is aide for enrolling Khata. 

Application for Khata: Get a Khata application structure (costs Rs. 10) or more duplicates of 5 archives (advancement charges receipt, Receipt of most recent Property Tax paid, building approval letters, title testament, representation of design). It is accessible either online or at any BBMP office. 

Application for Encumbrance testament (EC): Get an EC application (Form-22) accessible free of expense at sub-enlistment center office. It ought to be for a time of no less than a year. The expense of EC is Rs. 35 for one year. You should backpedal following a day or two to gather the EC. While applying for EC, one needs to convey duplicate of offers deed. 

Get Notarized duplicate of reports: Get Notarized of the accompanying Sale deed, Form II (on the off chance that you have), Possession Certificate (in the event that you have), GPA (if utilized), and top off the structure – green and pink. 

Present the Application: Tag the application to the authorized reports, EC, most recent property charge receipt, plan of the level alongside the normal archives and submit to the neighborhood ARO Office. Get the pink affirmation with accommodation number, stamped and marked by particular officer. 

Postliminary: Go back following 15 days to check the status. Continue doing this till you get the interest note. Interest note demonstrates the sum you have to pay as Khata enrollment expense. 

Evaluation of Property: Once you present the Khata application, the BBMP Revenue In-control and Assistant Revenue Officer by and by visits the property to survey the property. After the property is evaluated BBMP formally imparts this saying the property measurements (in sq feet), its quality according to BBMP appraisal and the expense risk subsequently. 

Pay enrollment charge: Submit the DDs according to the interest note to ARO Office and get a stamped affirmation. When you get affirmation, it implies the procedure has started. The organization charge or request note sum is two for every penny on the stamp paper esteem. A great many people confound this to be two for each penny of the property estimation said in the deal deed. The interest note is issued in bunches. So do co-ordinate with others in your building whose names are in the interest note to get every one of the DDs. 

Postliminary: Go back following 15 days to check the status. 

Khata warning: Check the notice to see that your name is spelt accurately and other data like range of level, auto park, thus on are right. Take the Khata notice to the ARO office with an application asking for issue of Khata testament (pay Rs. 25) and Khata concentrate (pay Rs. 100). 

Khata Registration: Once you pay the Khata Registration charge, in around 1-2 weeks, one gets the notification for paying the pending property charge. Without this the Khata Extract won't be issued in your name. However, in the event that you have come to this point it implies Khata has been in fact enrolled on your name. 

Subsequent is vital for effectively enlisting your Khata and getting your Khata Certificate. On the off chance that any BBMP official requests cash separated from the Khata enrollment charge and cost of use structure, you can hold up a grumbling with the concerned Revenue Officer. Try not to include center men for Khata enlistment. Methodology senior BBMP authorities at the zonal level on the off chance that you counter any issues. 

In the event that you don't get any reaction from BBMP following 2-3 months, record "Right to Information (RTI) application" to look for data on status of Khata enrollment. There is no recommended structure for application looking for data. Be that as it may, the application ought to have name and finish postal location of the candidate. The application can be made on plain paper. The candidate is not required to give any explanation behind looking for data. If there should be an occurrence of dismissal of your application, candidate would be educated the purposes behind dismissal

WE ARE PROVIDING HOMELOAN FOR B KHATA PROPERTY IN BANGALORE(EAST,NORTH,SOUNTH,WEST)

READ MORE:
www.bkhataloans.com